Fix structure to adhere to Gradle's convention
