+++ src/test/java/algorithms/graphs/breadthFirstSearchTest.java
@@ -2,7 +2,6 @@ package algorithms.graphs;
 
 import org.junit.Test;
 
-import algorithms.graphs.breadthFirstSearch;
 import algorithms.graphs.util.BinaryTreeNode;
 import algorithms.graphs.util.GraphNode;
