+++ src/test/java/dataStructures/hashSet/openAddressing/HashSetTest.java
@@ -1,10 +1,8 @@
 package dataStructures.hashSet.openAddressing;
 
 import org.junit.Test;
-import dataStructures.hashSet.openAddressing.HashSet;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.junit.Assert.*;
